requirements.txt
jj-wingman
gymnasium[mujoco]
